+++ inst/doc/univariate.R
@@ -4,10 +4,10 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-## ---- echo=FALSE, results='asis'----------------------------------------------
+## ----echo=FALSE, results='asis'-----------------------------------------------
 knitr::kable(head(survival::lung, 10))
 
-## ---- warning=FALSE, message=FALSE--------------------------------------------
+## ----warning=FALSE, message=FALSE---------------------------------------------
 library(tidyverse)
 library(tidytidbits)
 library(survivalAnalysis)
@@ -39,10 +39,10 @@ survival::lung %>%
                    by_label_map=c(`TRUE`="ECOG 1-2",
                                   `FALSE`="ECOG 2-3"))
 
-## ---- fig.width=6, fig.height=4-----------------------------------------------
+## ----fig.width=6, fig.height=4------------------------------------------------
 kaplan_meier_plot(result)
 
-## ---- fig.width=8, fig.height=5-----------------------------------------------
+## ----fig.width=8, fig.height=5------------------------------------------------
 kaplan_meier_plot(result,
                   break.time.by="breakByQuarterYear",
                   xlab=".OS.months",
@@ -52,7 +52,7 @@ kaplan_meier_plot(result,
                   table.layout="clean",
                   ggtheme=ggplot2::theme_bw(10))
 
-## ---- fig.width=8, fig.height=10, results='hide'------------------------------
+## ----fig.width=8, fig.height=10, results='hide'-------------------------------
 default_args <- list(break.time.by="breakByMonth",
                      xlab=".OS.months",
                      legend.title="ECOG Status",